+++ linstor_client/commands/rsc_dfn_cmds.py
@@ -196,6 +196,7 @@ class ResourceDefinitionCommands(Commands):
             description='Prints a list of all resource definitions known to '
             'LINSTOR. By default, the list is printed as a human readable table.')
         p_lrscdfs.add_argument('-p', '--pastable', action="store_true", help='Generate pastable output')
+        Commands.add_truncate_args(p_lrscdfs)
         p_lrscdfs.add_argument('-g', '--groupby', nargs='+',
                                choices=rsc_dfn_groupby,
                                type=str.lower).completer = rsc_dfn_group_completer
@@ -230,6 +231,7 @@ class ResourceDefinitionCommands(Commands):
             aliases=[Commands.Subcommands.ListProperties.SHORT],
             description="Prints all properties of the specified resource definitions.")
         p_sp.add_argument('-p', '--pastable', action="store_true", help='Generate pastable output')
+        Commands.add_truncate_args(p_sp)
         p_sp.add_argument(
             '--from-file',
             type=argparse.FileType('r'),
@@ -398,7 +400,8 @@ class ResourceDefinitionCommands(Commands):
         :param linstor.responses.ResourceDefinitionResponse lstmsg:
         :return:
         """
-        tbl = linstor_client.Table(utf8=not args.no_utf8, colors=not args.no_color, pastable=args.pastable)
+        tbl = linstor_client.Table(utf8=not args.no_utf8, colors=not args.no_color,
+                                   pastable=args.pastable, truncate=args.truncate)
 
         show_ext_name = args.external_name
         show_preferred_drbd_ports = args.show_preferred_drbd_ports